add point controller

+++ src/main/java/com/nhnacademy/taskapi/point/request/CreatePointPolicyRequest.java
@@ -3,13 +3,12 @@ package com.nhnacademy.taskapi.point.request;
 import com.nhnacademy.taskapi.point.domain.PointPolicy;
 import lombok.Builder;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Builder
-public record CreatePointPolicyRequest(String pointPolicyName, BigDecimal pointPolicyApply,
+public record CreatePointPolicyRequest(String pointPolicyName, int pointPolicyApply,
                                        String pointPolicyCondition, boolean pointPolicyApplyType,
-                                       BigDecimal pointPolicyConditionAmount) {
+                                       int pointPolicyConditionAmount) {
 
     public PointPolicy toEntity() {
         return PointPolicy.builder()
